# api/src/main/resources/swagger.yaml
swagger: '2.0'
info:
  title: To-Do List API
  version: 0.0.1
  description: This document contains the specification for the To-Do List API.
basePath: /todo-api/v1
tags:
  - name: todo
    description: Provides access to todo lists.
  - name: user
    description: Procides access to user management.
schemes:
  - http
paths:
  '/user/{userId}':
    get:
      tags:
        - user
      summary: Return a user.
      description: ''
      operationId: getUser
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          description: technical ID of the user
          required: true
          type: string
      responses:
        '200':
          description: Successful operation.
          schema:
            title: User
            type: object
            $ref: '#/definitions/User'
        '400':
          description: Bad Request - Invalid parameter value.
        '401':
          description: Unauthorized - No user credentials where provided.
        '404':
          description: Not Found - No result has been found matching the request criteria.
    post:
      tags:
        - "user"
      summary: Create a new user.
      description: ''
      operationId: createUser
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - name: userId
          in: path
          description: technical ID of the user
          required: true
          type: string
        - name: user
          in: body
          description: User to create.
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '201':
          description: Successful operation.
          schema:
            title: User
            type: object
            $ref: '#/definitions/User'
        '400':
          description: Bad Request - Invalid parameter value.
        '401':
          description: Unauthorized - No user credentials where provided.

  '/todo/{userId}':
    get:
      tags:
        - todo
      summary: Return the to do list of a user.
      description: ''
      operationId: getToDoList
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          description: technical ID of the user
          required: true
          type: string
      responses:
        '200':
          description: Successful operation.
          schema:
            title: A users To-Do List
            type: array
            items:
              $ref: '#/definitions/ToDoItem'
        '400':
          description: Bad Request - Invalid parameter value.
        '401':
          description: Unauthorized - No user credentials where provided.
        '404':
          description: Not Found - No result has been found matching the request criteria.
    post:
      tags:
        - todo
      summary: Create a new entry in the list
      description: ''
      operationId: addEntry
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          description: technical ID of the user
          required: true
          type: string
        - name: toDoItem
          in: body
          required: true
          schema:
            type: object
            $ref: '#/definitions/ToDoItem'
      responses:
        '200':
          description: Successful operation.
          schema:
            title: A users To-Do List
            type: object
            $ref: '#/definitions/ToDoItem'
        '400':
          description: Bad Request - Invalid parameter value.
        '401':
          description: Unauthorized - No user credentials where provided.
        '404':
          description: Not Found - No result has been found matching the request criteria.


definitions:
  User:
    type: object
    properties:
      userId:
        type: string
        description: technical user id
      name:
        type: string
        description: The user name
    xml:
      name: User
    example:
      userId: '0815'
      name: Hans Wurst
  ToDoItem:
    type: object
    properties:
      id:
        type: string
      title:
        type: string
      description:
        type: string
      done:
        type: boolean